% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukgovdatar-package.R
\docType{package}
\name{ukgovdatar-package}
\alias{ukgovdatar}
\alias{ukgovdatar-package}
\title{ukgovdatar: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petewrixon/ukgovdatar}
  \item Report bugs at \url{https://github.com/petewrixon/ukgovdatar/issues}
}

}
\keyword{internal}
